pyproject.toml
[tool.poetry]
name = "posting-auth-service"
version = "0.1.0"
description = "Auth service for Posting Micro-services"
authors = ["Luciano Gomes <lucianogvda02@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.112.1"}
pymongo = "^4.8.0"
mongoengine = "^0.28.2"
uvicorn = "^0.30.6"
pydantic = {extras = ["email"], version = "^2.8.2"}
pyjwt = "^2.9.0"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
python-multipart = "^0.0.9"
pydantic-settings = "^2.4.0"
testcontainers = "^4.8.0"
pytest = "^8.3.2"
freezegun = "^1.5.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.2"
taskipy = "^1.13.0"

[tool.taskipy.tasks]
lint = 'ruff check .; ruff check . --diff'
format = 'ruff check . --fix; ruff format .'
test = 'pytest -s -x -vv'

[tool.ruff]
line-length = 120
extend-exclude = ["migrations"]

[tool.ruff.lint]
preview = true
select = ["I", "F", "E", "W"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
